backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.0
python-dotenv==1.0.1
openai==1.54.0
httpx==0.27.2
aiosqlite==0.20.0

